Add files via upload
To run it, download the "  /Ants and Bugs+Twin CS135 Module 3 Project/Ants and Bugs+Twin CS135 Module 3 Project.sln "  file. It's the only .sln file and you will click on it to open Visual Studios and the project, but you need to put all these files into one folder so that Visual Studios can run it.